Fixed some docstrings. Split up IntentMaxHistoryTrackerFeaturizer into a separate module.
