+++ rasa/core/policies/unexpected_intent_policy.py
@@ -16,8 +16,8 @@ from rasa.shared.nlu.constants import (
     ACTION_NAME,
 )
 from rasa.nlu.extractors.extractor import EntityTagSpec
-from rasa.core.featurizers.tracker_featurizers import (
-    TrackerFeaturizer,
+from rasa.core.featurizers.tracker_featurizers import TrackerFeaturizer
+from rasa.core.featurizers.intent_max_history_tracker_featurizer import (
     IntentMaxHistoryTrackerFeaturizer,
 )
 from rasa.core.featurizers.single_state_featurizer import (
